+++ fiddles/c/fiddle-0018-CharArray/chararray/src/main.c
@@ -1,0 +1,11 @@
+#include <stdio.h>
+
+int main(void) {
+    char word[] = { 'H', 'e', 'l', 'l', 'o', '!' };
+    int i;
+    for ( i = 0; i < 6; ++i ) {
+        printf ( "%c", word[i] );
+    }
+    printf ("\n");
+    return 0;
+}